Animations!
I know this goes way beyond my commission, so feel free to revert
anything that you don't like, but I added a few animations to the TP
that I thought would go a long way towards making the map feel
immersive.  I added a subtle little glint to the 3 wool weapon blocks
to let the player know "do something with me."  I also added 8 item
animations: a glint to the 4 keys and the sapphire, blood dripping from
the bloody arrow, a little sparkle to the fire-starting wand, and in
honor of megga: a flopping raw fish.