More Animations
Added animations for TNT block and plant marker, and updated animations
for Raw Fish and Flint and Steel, including new raw fish .png file for
1.7.2. They're lots better.  Also changed the icon and added animations
for the Dwarven Ring of Power so it doesn't look like The One Ring
anymore.  (I assume it's not supposed to. Correct me if I'm wrong here)