*sparklez*
Repurposed Blue, Light Blue, and Magenta Stained Clay blocks to add
sparkle animations to the Saphhire Ore (Lapis Blocks) throughout the
map.  This can work in the veins of ore through the mines, but looks
especially awesome on the Alcar Sapphire.  If you like it, it will
require Dev or Bevs to go back and do a little editing.  Reference this
link to see what I'm talking about: http://youtu.be/o21RTPFN7Uw