Sounds
1) Changed the Zombie and Skelly sounds to sound like orcs.  Might need
some tweaking, but it’s scary as hell and plays well in-game.  2) Also
silenced the mine cart ride since it’s not really supposed to be a mine
cart. 3) Lastly, updated the lang files to change “Music Disc” to “Boss
Drop” since the last disc isn’t actually a disc.

Version incremented to 1.5